requirements.txt
langchain
langchain_community
langchainhub
langchain_huggingface
faiss-cpu
pypdf
flask
python-dotenv
